Modify exercises based on feedback
