+++ week_6/Ejercicio1.py
@@ -1,0 +1,14 @@
+"""
+💪🏽 **Ejercicios**
+1. Cree dos funciones que impriman dos cosas distintas,
+ y haga que la primera llame la segunda.
+"""
+
+def print_message_1():
+    print("Hello there")
+    print_message_2()  
+
+def print_message_2():
+    print("I am tested functions")
+
+print_message_1()
