--- week_9/archivo.txt
@@ -1,1 +1,0 @@
-https://github.com/Emanuel7TS/DUAD/pull/1
